% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_prefecture_code.r
\name{choose_prefecture_code}
\alias{choose_prefecture_code}
\title{Choose Prefecture Codes Interactively}
\usage{
choose_prefecture_code()
}
\description{
Display a popup window with prefectures checkboxes, and return the codes of checked prefectures.
}

