% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKSJData.R
\name{getKSJData}
\alias{getKSJData}
\title{Get JPGIS2.1 Data}
\usage{
getKSJData(zip_url, translate_columns = TRUE)
}
\arguments{
\item{zip_url}{The URL of the Zip file.}

\item{translate_columns}{If \code{TRUE}, try to use human-readable column names.}
}
\description{
Download and load spatial data from ZIP file downloaded from Kokudo Suuchi service. Note that this function does
not use API; directly download ZIP file and load the data by \link[rgdal]{readOGR}.
}
\examples{
\dontrun{
l <- getKSJData("http://nlftp.mlit.go.jp/ksj/gml/data/W07/W07-09/W07-09_3641-jgd_GML.zip")
names(l)
str(l, max.level = 1)
}
}
\seealso{
\url{http://nlftp.mlit.go.jp/ksj/api/about_api.html}
}

