% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKSJSummary.r
\name{getKSJSummary}
\alias{getKSJSummary}
\title{getKSJSummary API}
\usage{
getKSJSummary(appId = "ksjapibeta1", lang = "J", dataformat = 1)
}
\description{
getKSJSummary API
}
\seealso{
\url{http://nlftp.mlit.go.jp/ksj/api/about_api.html}
}

