% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKSJURL.r
\name{getKSJURL}
\alias{getKSJURL}
\title{getKSJURL API}
\usage{
getKSJURL(identifier, prefCode = NULL, meshCode = NULL, metroArea = NULL,
  fiscalyer = NULL, appId = "ksjapibeta1", lang = "J", dataformat = 1)
}
\description{
getKSJURL API
}
\seealso{
\url{http://nlftp.mlit.go.jp/ksj/api/about_api.html}
}

