% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kokudosuuchi-package.R
\docType{package}
\name{kokudosuuchi}
\alias{kokudosuuchi}
\alias{kokudosuuchi-package}
\title{Kokudo Suuchi Joho - Download Japanese National Land Numerical Information}
\description{
Use the Kokudo Suuchi Joho API
}
\seealso{
\url{http://nlftp.mlit.go.jp/ksj/api/about_api.html}
}

