% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KSJCodeDescriptionURL}
\alias{KSJCodeDescriptionURL}
\title{Corresponding Table Of Code Descriptions}
\format{A data.frame with codes and URLs of the description pages}
\source{
\url{http://nlftp.mlit.go.jp/ksj/index.html}
}
\usage{
KSJCodeDescriptionURL
}
\description{
Corresponding Table Of Code Descriptions
}
\keyword{datasets}
