% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KSJIdentifierDescriptionURL}
\alias{KSJIdentifierDescriptionURL}
\title{Corresponding Table Of Identifier Descriptions}
\format{A data.frame with identifier and URLs of the description pages}
\source{
\url{http://nlftp.mlit.go.jp/ksj/index.html}
}
\usage{
KSJIdentifierDescriptionURL
}
\description{
Corresponding Table Of Identifier Descriptions
}
\keyword{datasets}
