% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KSJMetadata_code}
\alias{KSJMetadata_code}
\alias{KSJShapeProperty}
\title{Corresponding Table Of Shapefile Properties}
\format{A data.frame with code and names}
\source{
\url{http://nlftp.mlit.go.jp/ksj/index.html}, \url{http://nlftp.mlit.go.jp/ksj/gml/shape_property_table.xls}

\url{http://nlftp.mlit.go.jp/ksj/gml/shape_property_table.xls}
}
\usage{
KSJMetadata_code

KSJShapeProperty
}
\description{
This data is deprecated. Use \link{KSJMetadata_code} instead.
}
\keyword{datasets}
