% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSJShapeProperties.R
\docType{data}
\name{KSJShapeProperty}
\alias{KSJShapeProperty}
\title{Corresponding Table Of Shapefile Properties}
\format{A data.frame with codes and names}
\source{
\url{http://nlftp.mlit.go.jp/ksj/gml/shape_property_table.xls}
}
\usage{
KSJShapeProperty
}
\description{
Corresponding Table Of Shapefile Properties
}
\keyword{datasets}
