% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj_data.R
\name{read_ksj_data}
\alias{read_ksj_data}
\title{Read JPGIS2.1 Data}
\usage{
read_ksj_data(zip_file, cache_dir = NULL, encoding = "CP932")
}
\arguments{
\item{zip_file}{A path to ZIP file downloaded from Kokudo Suuchi service.}

\item{cache_dir}{Path to a directory for caching zip files.}

\item{encoding}{Encoding of the data.}
}
\description{
Read JPGIS2.1 Data
}
