% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKSJData.R, R/translateKSJData.R
\name{readKSJData}
\alias{readKSJData}
\alias{translateKSJData}
\title{Read JPGIS2.1 Data}
\usage{
readKSJData(x, encoding = "CP932")

translateKSJData(x, quiet = TRUE)
}
\arguments{
\item{x}{Object of class \link[sf:sf]{sf}}

\item{encoding}{Encoding of the data.}

\item{quiet}{If \code{TRUE}, suppress messages.}
}
\description{
\code{readKSJData()} is an utility to read data downloaded from 'Kokudo Suuchi' service.
}
\examples{
\dontrun{
zip_file <- tempfile(fileext = ".zip")
url <- "https://nlftp.mlit.go.jp/ksj/gml/data/W07/W07-09/W07-09_3641-jgd_GML.zip"
download.file(url, zip_file)
readKSJData(zip_file)
}

}
