% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_in_peace.R
\name{getKSJSummary}
\alias{getKSJSummary}
\alias{getKSJURL}
\alias{choose_prefecture_code}
\title{Rest in peace, 'Kokudo Suuchi' API...}
\usage{
getKSJSummary(...)

getKSJURL(...)

choose_prefecture_code(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Rest in peace, 'Kokudo Suuchi' API...
}
\keyword{internal}
