% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKSJData.R
\name{getKSJData}
\alias{getKSJData}
\title{Get JPGIS2.1 Data}
\usage{
getKSJData(zip_file, cache_dir = tempdir(), encoding = "CP932")
}
\arguments{
\item{zip_file}{Either a URL, a path to a zip file, or a path to a directory which contains shape files.}

\item{cache_dir}{Path to a directory for caching zip files.}

\item{encoding}{Encoding of the data.}
}
\description{
\code{getKSJData()} is deprecated.
}
\keyword{internal}
