% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KSJMetadata_code}
\alias{KSJMetadata_code}
\alias{KSJMetadata_code_correspondence_tables}
\alias{KSJMetadata_code_year_cols}
\title{Corresponding Table Of Shapefile Properties}
\format{A data.frame with code and names}
\source{
\url{http://nlftp.mlit.go.jp/ksj/index.html}, \url{http://nlftp.mlit.go.jp/ksj/gml/shape_property_table.xls}

\url{http://nlftp.mlit.go.jp/ksj/index.html}

\url{http://nlftp.mlit.go.jp/ksj/index.html}
}
\usage{
KSJMetadata_code

KSJMetadata_code_correspondence_tables

KSJMetadata_code_year_cols
}
\description{
Corresponding Table Of Shapefile Properties
}
\keyword{datasets}
