% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSJprefCodes.r
\docType{data}
\name{KSJprefCodes}
\alias{KSJprefCodes}
\title{Corresponding Table Of Names And Codes Of Prefectures}
\format{A data.frame with prefecture names and codes}
\source{
\url{http://nlftp.mlit.go.jp/ksj/gml/codelist/PrefCd.html}
}
\usage{
KSJprefCodes
}
\description{
Corresponding Table Of Names And Codes Of Prefectures
}
\keyword{datasets}

