% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSJPrefCodes.R
\docType{data}
\name{KSJPrefCodes}
\alias{KSJPrefCodes}
\title{Corresponding Table Of Names And Codes Of Prefectures}
\format{A data.frame with prefecture names and codes}
\source{
\url{http://nlftp.mlit.go.jp/ksj/gml/codelist/PrefCd.html}
}
\usage{
KSJPrefCodes
}
\description{
Corresponding Table Of Names And Codes Of Prefectures
}
\keyword{datasets}
