% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKSJData.R, R/translateKSJData.R
\name{readKSJData}
\alias{readKSJData}
\alias{translateKSJData}
\title{Read JPGIS2.1 Data}
\usage{
readKSJData(x, encoding = "CP932")

translateKSJData(
  x,
  id = NULL,
  variant = NULL,
  quiet = TRUE,
  translate_colnames = TRUE,
  translate_codelist = TRUE
)
}
\arguments{
\item{x}{A list of \link[sf:sf]{sf} objects.}

\item{encoding}{Encoding of the data.}

\item{id}{An ID of the dataset (e.g. \code{A03}). This can be automatically detected if \code{x}
is loaded by \code{\link[=readKSJData]{readKSJData()}}.}

\item{variant}{A type of variant in case the translation cannot be determined only by \code{id}.}

\item{quiet}{If \code{TRUE}, suppress messages.}

\item{translate_colnames}{If \code{TRUE}, translate colnames to human readable labels.}

\item{translate_codelist}{If \code{TRUE}, translate codes to human readable labels.}
}
\description{
\code{readKSJData()} is an utility to read data downloaded from 'Kokudo Suuchi' service.
}
\examples{
\dontrun{
zip_file <- tempfile(fileext = ".zip")
url <- "https://nlftp.mlit.go.jp/ksj/gml/data/W07/W07-09/W07-09_3641-jgd_GML.zip"
download.file(url, zip_file)
readKSJData(zip_file)
}

}
